# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: DishFlow API
  version: 0.1.0
  description: |
    DishFlow 餐饮门店点餐与履约系统 API 契约（PRD §16）。

    ## 约定
    - 业务 API 前缀 `/api/v1`；JSON 使用 snake_case。
    - 成功直接返回业务对象，或 `{items,next_cursor,total}` 列表分页。
    - 错误统一为 `{code,message,request_id,details}`。
    - 管理端使用 Cookie + `X-Store-Id`；小程序使用 Bearer token + `X-Wechat-Appid`。
    - 所有关键写接口接受 `Idempotency-Key`。
    - 金额一律 int64 分；时间使用 RFC3339。
  contact:
    name: DishFlow
servers:
  - url: http://127.0.0.1:8080
    description: 本地开发
...
paths:
...
  /api/v1/store:
    get:
      tags: [Storefront]
      operationId: storeInfo
      security: [{ wechatAppidHeader: [] }]
      responses:
        '200': { description: 门店信息, content: { application/json: { schema: { type: object } } } }
...
  /api/v1/pickup-slots:
    get:
      tags: [Storefront]
      operationId: pickupSlots
      security: [{ wechatAppidHeader: [] }]
      parameters:
        - { name: date, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: 预约时段, content: { application/json: { schema: { type: object } } } }
  /api/v1/tables/resolve:
    get:
      tags: [Storefront]
      operationId: resolveTable
      security: [{ wechatAppidHeader: [] }]
      parameters:
        - { name: token, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: 桌台, content: { application/json: { schema: { type: object } } } }
        '404': { description: 桌码无效, content: { application/json: { schema: { $ref: '#/components/schemas/Error', example: { code: TABLE_NOT_FOUND, message: table not found, request_id: req_1 } } } } }
  /api/v1/auth/wechat/session:
    post:
      tags: [Storefront]
      operationId: wechatSession
      security: [{ wechatAppidHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { code: { type: string } }, required: [code] } } } }
      responses:
        '200': { description: 会话, content: { application/json: { schema: { type: object, properties: { token: { type: string }, expires_at: { type: string, format: date-time } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/coupon-offers:
    get:
      tags: [Customer]
      operationId: couponOffers
      security: [{ wechatAppidHeader: [] }]
      responses:
        '200': { description: 可领券, content: { application/json: { schema: { $ref: '#/components/schemas/List' } } } }
...
components:
  securitySchemes:
    customerBearer:
      type: http
      scheme: bearer
    adminCookie:
      type: apiKey
      in: cookie
      name: shop_session
    storeIdHeader:
      type: apiKey
      in: header
      name: X-Store-Id
    wechatAppidHeader:
      type: apiKey
      in: header
      name: X-Wechat-Appid
...
  responses:
...
    Unauthorized: { description: 未认证, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          enum: [INTERNAL, BAD_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, STATE_CONFLICT, RATE_LIMITED, QUOTE_EXPIRED, QUOTE_MISMATCH, TABLE_NOT_FOUND, TABLE_DISABLED, PICKUP_SLOT_FULL, PICKUP_TIME_INVALID, PAYMENT_UNAVAILABLE, REFUND_CONFLICT, INSUFFICIENT_POINTS, WECHAT_APPID_CONFLICT]
        message: { type: string }
        request_id: { type: string }
        details: {}
    List:
      type: object
      required: [items]
      properties:
        items: { type: array, items: {} }
        next_cursor: { type: string }
        total: { type: integer, format: int64 }
...
security: []